# pandedaily/server/src/docs/swagger.docs.yaml
openapi: 3.0.0

info:
  title: PandeDaily Control Panel API
  description: API for PandeDaily Admin Dashboard
  version: 0.0.1

servers:
  - url: http://localhost:3080
    description: Local development server

# Remove global security if endpoints don't require auth
# Or specify which endpoints require auth
# security:
#   - SessionAuth: []
#   - BearerAuth: []

tags:
  - name: Authentication
    description: Authentication endpoints

paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user account
      operationId: login
      # Remove security for login endpoint
      # security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
            example: # Add example for Try it out
              username: 'admin'
              password: 'password123'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example: # Add response example
                message: 'Login successful'
                token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
                user:
                  id: 1
                  fullname: 'John Doe'
                  access_id: 1
                  username: 'admin'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Invalid username or password'
        '500':
          $ref: '#/components/responses/Error500'

  /auth/check:
    get:
      tags: [Authentication]
      summary: Check current session
      operationId: checkSession
      # This endpoint likely requires authentication
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Session valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 1
                fullname: 'John Doe'
                access_id: 1
                username: 'admin'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/Error500'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      operationId: logout
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/Error500'

components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Login:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password123

    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        fullname:
          type: string
          example: John Doe
        access_id:
          type: integer
          example: 1
        username:
          type: string
          example: admin

    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong

  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Unauthorized access'

    Error500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Internal server error'